Move Unit and UnitStats to new file unit.ts from game.ts
Also rename buildings.ts as building.ts
