+++ Server/types.ts
@@ -69,6 +69,7 @@ export type Coordinate = [number, number];
 export type SocketEvent = {event: string, params: any}
 
 export type PlayerArr<T> = [T[], T[]];
+export const emptyPArr = <T>() => [[[], []], [[], []]] as PlayerArr<T[]>;
 
 export type ClientGameState = {
     turn: number,
